changed enemy behaviour, added groundwork for bullets and player movement

+++ phase1/app/src/main/java/com/example/userinterface/GameManager/SpaceInvaders/SpaceObject.java
@@ -20,7 +20,7 @@ abstract class SpaceObject {
 
     private double speed;
 
-    SpaceObject(int x, int y) {
+    SpaceObject(float x, float y) {
         this.setX(x);
         this.setY(y);
     }
